# Fix limitations of the parse state merging algorithm to produce tables with fewer states